# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Simulador Hipotecas Api Rest",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/",
            "description": "Localhost env"
        }
    ],
    "paths": {
        "/clientes": {
            "get": {
                "tags": [
                    "Operaciones relacionadas con clientes"
                ],
                "summary": "Obtener dnis de clientes",
                "description": "Devuelve los dnis de los clientes existentes",
                "operationId": "index",
                "responses": {
                    "200": {
                        "description": "Clientes obtenidos con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "number",
                                            "example": "200"
                                        },
                                        "response": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "ID_cliente": {
                                                        "type": "number"
                                                    },
                                                    "dni": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No hay clientes registrados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "number",
                                            "example": "404"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "No hay clientes registrados"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Operaciones relacionadas con clientes"
                ],
                "summary": "Creación de un nuevo cliente",
                "description": "Creación de un nuevo cliente no existente en la base de datos",
                "operationId": "store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "dni",
                                    "email",
                                    "capital_solicitado"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Pepe"
                                    },
                                    "dni": {
                                        "type": "string",
                                        "format": "dni",
                                        "uniqueItems": true,
                                        "example": "12345678Z"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "pepe@gmail.com"
                                    },
                                    "capital_solicitado": {
                                        "type": "number",
                                        "format": "float",
                                        "minimum": 0,
                                        "example": "200000"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cliente creado con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "number",
                                            "example": "201"
                                        },
                                        "response": {
                                            "properties": {
                                                "nombre": {
                                                    "type": "string",
                                                    "example": "Pepe"
                                                },
                                                "dni": {
                                                    "type": "string",
                                                    "example": "12345678Z"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "pepe@gmail.com"
                                                },
                                                "capital_solicitado": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": "200000"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Contenido Inprocesable",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The x field {error type}"
                                        },
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "error_field": {
                                                        "example": "The x field {error type}"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error Interno"
                    }
                }
            }
        },
...
    }
}